+++ src/module.ts
@@ -2,9 +2,10 @@ import { DataSourcePlugin } from '@grafana/data';
 import { DataSource } from './datasource';
 import { ConfigEditor } from './components/ConfigEditor';
 import { QueryEditor } from './components/QueryEditor';
+import { VariableQueryEditor } from './components/VariableQueryEditor';
 import { CatalystQuery, CatalystJsonData } from './types';
 
 export const plugin = new DataSourcePlugin<DataSource, CatalystQuery, CatalystJsonData>(DataSource)
   .setConfigEditor(ConfigEditor)
-  .setQueryEditor(QueryEditor);
+  .setQueryEditor(QueryEditor)
   .setVariableQueryEditor(VariableQueryEditor);